Game Over
Created a game over overlay and function to display final results and outcome. Also, fixed click event issue on the choices container that would register a click on the empty space of the ordered list. Last step is to let the user start a new game without refreshing the page.